+++ examples/simple_chain/lib/simple_chain.ex
@@ -1,0 +1,18 @@
+defmodule SimpleChain do
+  @moduledoc """
+  Documentation for SimpleChain.
+  """
+
+  @doc """
+  Hello world.
+
+  ## Examples
+
+      iex> SimpleChain.hello
+      :world
+
+  """
+  def hello do
+    :world
+  end
+end